chore: Remove redundant clean_requirements function and update package_utils
